included the geo_sysdef/cm_sysdef method option
